TD in Field should be float

+++ db/migrate/20200903203603_add_columns_to_field.rb
@@ -1,6 +1,6 @@
 class AddColumnsToField < ActiveRecord::Migration[5.2]
   def change
-    add_column :fields, :depth, :integer
+    add_column :fields, :depth, :float
     add_column :fields, :tile_bioreactors, :boolean
     add_column :fields, :drainage_water_management, :boolean
   end
